+++ packages/spacecat-shared-utils/src/index.d.ts
@@ -43,6 +43,8 @@ export function toBoolean(value: unknown): boolean;
 
 export function isValidUrl(urlString: string): boolean;
 
+export function isValidHelixPreviewUrl(urlString: string): boolean;
+
 export function isValidUUID(uuid: string): boolean;
 
 export function isValidIMSOrgId(imsOrgId: string): boolean;